+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Added
 
+- Transfer focus management and keyboard navigation to the tiles engine @sneridagh
+
 ### Changes
 
 - Slight amendment to Blocks chooser styling @sneridagh
+- The default view for content types `DocumentView.jsx` has been renamed to a more appropiate `DefaultView.jsx` @sneridagh
 
 ### Internal